+++ app/src/main/java/org/ole/planet/myplanet/di/RepositoryModule.kt
@@ -23,6 +23,8 @@ import org.ole.planet.myplanet.repository.DownloadRepository
 import org.ole.planet.myplanet.repository.DownloadRepositoryImpl
 import org.ole.planet.myplanet.repository.EventsRepository
 import org.ole.planet.myplanet.repository.EventsRepositoryImpl
+import org.ole.planet.myplanet.repository.EnterprisesRepository
+import org.ole.planet.myplanet.repository.EnterprisesRepositoryImpl
 import org.ole.planet.myplanet.repository.EventsSyncWriter
 import org.ole.planet.myplanet.repository.FeedbackRepository
 import org.ole.planet.myplanet.repository.FeedbackRepositoryImpl
@@ -105,6 +107,10 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindEventsRepository(impl: EventsRepositoryImpl): EventsRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindEnterprisesRepository(impl: EnterprisesRepositoryImpl): EnterprisesRepository
 
     @Binds
     @Singleton
