+++ app/src/main/java/org/ole/planet/myplanet/repository/TeamsRepositoryImpl.kt
@@ -517,34 +517,6 @@ class TeamsRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun addReport(report: FinanceReportParams) {
-        val reportId = UUID.randomUUID().toString()
-        val doc = JsonObject().apply {
-            addProperty("_id", reportId)
-            addProperty("createdDate", timeProvider.now())
-            addProperty("description", report.description)
-            addProperty("beginningBalance", report.beginningBalance)
-            addProperty("sales", report.sales)
-            addProperty("otherIncome", report.otherIncome)
-            addProperty("wages", report.wages)
-            addProperty("otherExpenses", report.otherExpenses)
-            addProperty("startDate", report.startDate)
-            addProperty("endDate", report.endDate)
-            addProperty("updatedDate", timeProvider.now())
-            addProperty("teamId", report.teamId)
-            addProperty("teamType", report.teamType)
-            addProperty("teamPlanetCode", report.teamPlanetCode)
-            addProperty("docType", "report")
-            addProperty("updated", true)
-        }
-        val reportEntry = MyTeam().apply { _id = reportId }
-        MyTeam.populateTeamFields(doc, reportEntry)
-        teamDao.upsert(reportEntry.requireRoomEntity())
-        if (report.imageName != null && report.imageData != null) {
-            attachTeamImage(reportId, report.imageName, report.imageData)
-        }
-    }
-
     private suspend fun attachTeamImage(teamId: String, imageName: String, imageData: ByteArray) {
         if (teamId.isBlank()) return
         val destFile = MyTeam.getAttachmentFile(MainApplication.context, teamId, imageName) ?: return
@@ -555,40 +527,6 @@ class TeamsRepositoryImpl @Inject constructor(
         updateTeamEntityById(teamId) { team ->
             team.imageName = imageName
             team.updated = true
-        }
-    }
-
-    override suspend fun updateReport(reportId: String, payload: FinanceReportParams) {
-        if (reportId.isBlank()) return
-        val doc = JsonObject().apply {
-            addProperty("description", payload.description)
-            addProperty("beginningBalance", payload.beginningBalance)
-            addProperty("sales", payload.sales)
-            addProperty("otherIncome", payload.otherIncome)
-            addProperty("wages", payload.wages)
-            addProperty("otherExpenses", payload.otherExpenses)
-            addProperty("startDate", payload.startDate)
-            addProperty("endDate", payload.endDate)
-            addProperty("updatedDate", timeProvider.now())
-            addProperty("updated", true)
-        }
-        updateTeamEntityById(reportId) { report ->
-            MyTeam.populateReportFields(doc, report)
-            report.updated = true
-            if (report.updatedDate == 0L) {
-                report.updatedDate = timeProvider.now()
-            }
-        }
-        if (payload.imageName != null && payload.imageData != null) {
-            attachTeamImage(reportId, payload.imageName, payload.imageData)
-        }
-    }
-
-    override suspend fun archiveReport(reportId: String) {
-        if (reportId.isBlank()) return
-        updateTeamEntityById(reportId) { report ->
-            report.status = "archived"
-            report.updated = true
         }
     }
 
@@ -782,41 +720,6 @@ class TeamsRepositoryImpl @Inject constructor(
 
     override suspend fun getTasksByTeamId(teamId: String): Flow<List<TeamTask>> {
         return teamTaskDao.getTasksByTeamId(teamId)
-    }
-
-    override suspend fun getReportsFlow(teamId: String): Flow<List<MyTeam>> {
-        return teamDao.observeAll().map { entities ->
-            entities.filter {
-                it.teamId == teamId &&
-                    it.docType == "report" &&
-                    it.status != "archived"
-            }.sortedByDescending { it.createdDate }
-                .map { it }
-        }
-    }
-
-    override suspend fun exportReportsAsCsv(reports: List<MyTeam>, teamName: String): String {
-        val csvBuilder = StringBuilder()
-        csvBuilder.append(teamName).append(" Financial Report Summary\n\n")
-        csvBuilder.append("Start Date, End Date, Created Date, Updated Date, Beginning Balance, Sales, Other Income, Wages, Other Expenses, Profit/Loss, Ending Balance\n")
-        for (report in reports) {
-            val totalIncome = report.sales + report.otherIncome
-            val totalExpenses = report.wages + report.otherExpenses
-            val profitLoss = totalIncome - totalExpenses
-            val endingBalance = profitLoss + report.beginningBalance
-            csvBuilder.append(TimeUtils.formatDateForCsv(report.startDate)).append(", ")
-                .append(TimeUtils.formatDateForCsv(report.endDate)).append(", ")
-                .append(TimeUtils.formatDateForCsv(report.createdDate)).append(", ")
-                .append(TimeUtils.formatDateForCsv(report.updatedDate)).append(", ")
-                .append(report.beginningBalance).append(", ")
-                .append(report.sales).append(", ")
-                .append(report.otherIncome).append(", ")
-                .append(report.wages).append(", ")
-                .append(report.otherExpenses).append(", ")
-                .append(profitLoss).append(", ")
-                .append(endingBalance).append('\n')
-        }
-        return csvBuilder.toString()
     }
 
     override suspend fun deleteTask(taskId: String) {
